# onux/apps/user-service/openapi.yaml
openapi: 3.1.0
info:
  title: Cinacoin User Service API
  description: |
    Microservice for user and team management in the Cinacoin platform.
    Handles user profiles, team creation, team membership, and permissions.
  version: 1.0.0
  contact:
    name: Cinacoin Engineering
    url: https://cinacoin.com
  license:
    name: MIT

servers:
  - url: http://localhost:3201
    description: Local development
  - url: https://user-service.cinacoin.com
    description: Production

tags:
  - name: Users
    description: User management endpoints
  - name: Teams
    description: Team management endpoints
  - name: Team Members
    description: Team membership management
  - name: Health
    description: Service health checks

paths:
  /api/health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is degraded

  /api/users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicUser'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Users]
      summary: Create user (service auth only)
      operationId: createUser
      security:
        - ServiceApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: User already exists
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicUser'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Users]
      summary: Delete user (soft delete)
      operationId: deleteUser
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/users/{id}/teams:
    get:
      tags: [Users, Teams]
      summary: Get user's teams
      operationId: getUserTeams
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicTeam'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/teams:
    get:
      tags: [Teams]
      summary: List teams
      operationId: listTeams
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicTeam'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Teams]
      summary: Create team
      operationId: createTeam
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicTeam'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Team slug already exists

  /api/teams/{id}:
    get:
      tags: [Teams]
      summary: Get team by ID
      operationId: getTeamById
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicTeam'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Teams]
      summary: Update team
      operationId: updateTeam
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicTeam'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/teams/{id}/members:
    get:
      tags: [Team Members]
      summary: Get team members
      operationId: getTeamMembers
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicTeamMember'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Team Members]
      summary: Add team member
      operationId: addTeamMember
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTeamMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicTeamMember'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/teams/{id}/members/{userId}:
    put:
      tags: [Team Members]
      summary: Update team member role
      operationId: updateTeamMemberRole
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PublicTeamMember'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Team Members]
      summary: Remove team member
      operationId: removeTeamMember
      security:
        - BearerAuth: []
        - ServiceApiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT token from auth-service
    ServiceApiKey:
      type: apiKey
      in: header
      name: X-Service-API-Key
      description: Service-to-service API key

  schemas:
    UserStatus:
      type: string
      enum: [pending, active, suspended, disabled, deleted]

    TeamMemberRole:
      type: string
      enum: [owner, admin, member, viewer]

    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        displayName:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        locale:
          type: string
        timezone:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, externalId, email, emailVerified, status, locale, timezone, createdAt, updatedAt]

    PublicTeam:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdBy:
          type: string
          format: uuid
          nullable: true
        isActive:
          type: boolean
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, externalId, name, slug, isActive, createdAt, updatedAt]

    PublicTeamMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/TeamMemberRole'
        isActive:
          type: boolean
        invitedBy:
          type: string
          format: uuid
          nullable: true
        joinedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, teamId, userId, role, isActive, joinedAt, createdAt, updatedAt]

    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [total, page, pageSize, totalPages]

    CreateUserRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 100
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        avatarUrl:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/UserStatus'
        metadata:
          type: object

    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 100
        firstName:
          type: string
          maxLength: 50
        lastName:
          type: string
          maxLength: 50
        avatarUrl:
          type: string
          format: uri
        locale:
          type: string
          maxLength: 10
        timezone:
          type: string
          maxLength: 50
        status:
          $ref: '#/components/schemas/UserStatus'
        emailVerified:
          type: boolean
        metadata:
          type: object

    CreateTeamRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
        metadata:
          type: object

    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 50
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
          maxLength: 500
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        metadata:
          type: object

    AddTeamMemberRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/TeamMemberRole'

    UpdateMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/TeamMemberRole'

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, degraded]
        timestamp:
          type: string
          format: date-time
        service:
          type: string
        version:
          type: string
        database:
          type: string
          enum: [connected, disconnected]
      required: [status, timestamp, service]

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
      required: [error, message]

  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
            required: [error, details]